Exporting

For added convenience, InformationGeometry.jl already provides several methods, which can be used to export results like confidence regions or geodesics.

InformationGeometry.SaveConfidence — Function

SaveConfidence(sols::AbstractVector{<:AbstractODESolution}, N::Int=500; sigdigits::Int=7, adaptive::Bool=true) -> Matrix
SaveConfidence(Planes::AbstractVector{<:Plane}, sols::AbstractVector{<:AbstractODESolution}, N::Int=500; sigdigits::Int=7, adaptive::Bool=true) -> Matrix

Returns a Matrix of with N rows corresponding to the number of evaluations of each ODESolution in sols. The colums correspond to the various components of the evaluated solutions. E.g. for an ODESolution with 3 components, the 4. column in the Matrix corresponds to the evaluated first components of sols[2].

In particular, choosing the keyword adaptive=true samples the ODESolution objects roughly proportional to their curvature (instead of equidistant), which means that more samples are provided from tight bends than from segments that are straight, leading to more faithful representations of the confidence boundary when plotting.

InformationGeometry.SaveDataSet — Function

SaveDataSet(DS::DataSet; sigdigits::Int=0)

Returns a DataFrame whose columns respectively constitute the x-values, y-values and standard distributions associated with the data points. For sigdigits > 0 the values are rounded to the specified number of significant digits.