Providing Datasets

Typically, one of the most difficult parts of any data science problem is to bring the data into a form which lends itself to the subsequent analysis. This section aims to describe the containers used by InformationGeometry.jl to store datasets and models in detail.

The data itself is stored using the DataSet container.

InformationGeometry.DataSet — Type

The DataSet type is a versatile container for storing data. Typically, it is constructed by passing it three vectors x, y, sigma where the components of sigma quantify the standard deviation associated with each y-value. Alternatively, a full covariance matrix can be supplied for the ydata instead of a vector of standard deviations. The contents of a DataSet DS can later be accessed via xdata(DS), ydata(DS), sigma(DS).

Examples:

In the simplest case, where all data points are mutually independent and have a single $x$-component and a single $y$-component each, a DataSet consisting of four points can be constructed via

DataSet([1,2,3,4],[4,5,6.5,7.8],[0.5,0.45,0.6,0.8])

or alternatively by

DataSet([1,2,3,4],[4,5,6.5,7.8],Diagonal([0.5,0.45,0.6,0.8].^2))

where the diagonal covariance matrix in the second line is equivalent to the vector of uncertainties supplied in the first line.

More generally, if a dataset consists of $N$ points where each $x$-value has $n$ many components and each $y$-value has $m$ many components, this can be specified to the DataSet constructor via a tuple $(N,n,m)$ in addition to the vectors x, y and the covariance matrix. For example:

X = [0.9, 1.0, 1.1, 1.9, 2.0, 2.1, 2.9, 3.0, 3.1, 3.9, 4.0, 4.1]
Y = [1.0, 5.0, 4.0, 8.0, 9.0, 13.0, 16.0, 20.0]
Cov = Diagonal([2.0, 4.0, 2.0, 4.0, 2.0, 4.0, 2.0, 4.0])
dims = Tuple([4,3,2])
DS = DataSet(X,Y,Cov,dims)

To complete the specification of an inference problem, a model function which is assumed to model the relationship which is inherent in the takes an x-value and a parameter configuration $\theta$ must be added.

InformationGeometry.DataModel — Type

In addition to storing a DataSet, a DataModel also contains a function model(x,θ) and its derivative dmodel(x,θ) where x denotes the x-value of the data and θ is a vector of parameters on which the model depends. Crucially, dmodel contains the derivatives of the model with respect to the parameters θ, not the x-values. For example

DataSet([1,2,3,4],[4,5,6.5,7.8],[0.5,0.45,0.6,0.8])
model(x,θ::Vector) = θ[1] .* x .+ θ[2]
DM = DataModel(DS,model)

In order for all methods to perform as expected, the output of the model must always be in the form of a vector of numbers (if the y-values have more than one component or the model is evaluated on more than one point).

If a DataModel is constructed as shown above, the gradient of the model with respect to the parameters θ (i.e. its "Jacobian") will be calculated using automatic differentiation. Alternatively, an explicit analytic expression for the Jacobian can be specified by hand:

function dmodel(x,θ::Vector)
   J = Array{Float64}(undef, length(x), length(θ))
   @. J[:,1] = x        # ∂(model)/∂θ₁
   @. J[:,2] = 1.       # ∂(model)/∂θ₂
   return J
end
DM = DataModel(DS,model,dmodel)

The output of the Jacobian must be a matrix whose columns correspond to the partial derivatives with respect to different components of θ and whose rows correspond to evaluations at different values of x.

The DataSet contained in a DataModel DM can be accessed via DM.Data, whereas the model and its Jacobian can be used via DM.model and DM.dmodel respectively.